==> 01_cities_weather_infos.ipynb <==
# Usuals imports 
import numpy as np
import pandas as pd 
import requests
from time import sleep

## Cities 

cities = ["Mont Saint Michel", "St Malo", "Bayeux", "Le Havre", "Rouen", "Paris", "Amiens", "Lille", "Strasbourg", "Chateau du Haut Koenigsbourg", "Colmar", "Eguisheim", "Besancon", "Dijon", "Annecy", "Grenoble", "Lyon", "Gorges du Verdon", "Bormes les Mimosas", "Cassis", "Marseille", "Aix en Provence", "Avignon", "Uzes", "Nimes", "Aigues Mortes", "Saintes Maries de la mer", "Collioure", "Carcassonne", "Ariege", "Toulouse", "Montauban", "Biarritz", "Bayonne", "La Rochelle"]

## Dataframe Cities 

cities_locations = pd.DataFrame(columns= ['city', 'country', 'latitude', 'longitude', 'adress'])

for city in cities:
    country = 'France'
    latitude = 0.0
    longitude = 0.0
    adress = ''

    # Creating urls params
    payload = {'city': city, 'country': country, 'format': 'jsonv2'}

    # Create the requests 
    r = requests.get('https://nominatim.openstreetmap.org/search.php?', params=payload)

    # check result available 
    if r.status_code == 200 :
        r_result = pd.DataFrame(r.json())
        #attrib location 
        try: 
            latitude = r_result['lat'][0]
            longitude = r_result['lon'][0]
            adress =  r_result['display_name'][0]

        except:
            #try with county 
            payload = {'county': city, 'country': country, 'format': 'jsonv2'}
            r = requests.get('https://nominatim.openstreetmap.org/search.php?', params=payload)
            r_result = pd.DataFrame(r.json())

            try: 
                latitude = r_result['lat'][0]
                longitude = r_result['lon'][0]
                adress =  r_result['display_name'][0]

            except:
                #try with street 
                payload = {'street': city, 'country': country, 'format': 'jsonv2'}
                r = requests.get('https://nominatim.openstreetmap.org/search.php?', params=payload)
                r_result = pd.DataFrame(r.json())

                try: 
                    latitude = r_result['lat'][0]
                    longitude = r_result['lon'][0]
                    adress =  r_result['display_name'][0]
                
                except:
                    pass

        new_city = pd.DataFrame(columns= ['city', 'country', 'latitude', 'longitude', 'adress'], 
                                   data= [[ city, country, latitude, longitude, adress]])

        cities_locations = pd.concat([cities_locations, new_city])

# reset index due to concat
cities_locations.reset_index(inplace= True)
cities_locations.drop(columns=['index'], inplace= True)

# copy information in adress
cities_locations.drop(columns=['country'], inplace= True)

## Dataframe Wheater in Cities :

# Hide SettingWithCopyWarning
pd.options.mode.chained_assignment = None

KEYS = open('src/KEYS.txt', 'r').read()

# Init dataframe
cities_weather_infos = cities_locations.assign(weather_code = 0, felt_temperature=0.0, cloud_coverage=0, precipitation=0.0, wind=0.0)

for item in range(len(cities_weather_infos)):
    latitude = cities_weather_infos['latitude'][item]
    longitude = cities_weather_infos['longitude'][item]

    # Creating urls params
    payload = {'key': KEYS, 'lat': latitude, 'lon': longitude, }

    # carefull to limit request (60/secondes) for this API
    sleep(1)

    # Create the requests 
    r = requests.get('http://api.weatherbit.io/v2.0/current?', params=payload)

    # check result available 
    if r.status_code == 200 :
        r_result = pd.DataFrame(r.json())
        data = r_result['data'][0]
        weather = data['weather']

        try:
            # Somes interrestings values, more in https://www.weatherbit.io/api/weather-current 
            pressure = data['pres'] # Pressure (mb)
            date = data['ob_time'] # Date datetime
            temperature = data['temp'] # Temperature (c°)
            felt_temperature = data['app_temp'] # Apparent temperature (c°)
            humidity = data['rh'] # Relative humidity (%)
            cloud_coverage = data['clouds'] # Cloud coverage (%)
            dew_point = data['dewpt'] # Dew point (c°)
            wind = data['wind_spd'] # Wind speed (m/s)
            precipitation = data['precip'] # Precipitation 
            indice_uv = data['uv'] # Mean as 3 to 5
            weather_code = weather['code'] # Weather code, info in https://www.weatherbit.io/api/codes
            description = weather['description'] # Weather description

            # Keep usefull values
            cities_weather_infos['weather_code'][item] = weather_code
            cities_weather_infos['felt_temperature'][item] = felt_temperature
            cities_weather_infos['cloud_coverage'][item] = cloud_coverage
            cities_weather_infos['precipitation'][item] = precipitation
            cities_weather_infos['wind'][item] = wind

        except:
            pass


cities_weather_infos.to_csv(r'src/cities_weather_infos.csv', index=False)
