==> 07_hostels_list_located.ipynb <==
# Usuals imports 
import numpy as np
import pandas as pd 
import requests
from time import sleep
import re
import urllib.parse

## Import hostels list

hostels_list = pd.read_json('src/hostels_list.json', encoding='utf-8')

def html_unicorn(sentence):
    foo = sentence.replace("\\/", "/").encode().decode('utf-8')
    foo = foo.replace(f'\n', '')
    return foo

for index in range(len(hostels_list)):
    if hostels_list['hotel_name'][index] == None:
        hostel_url = hostels_list['booking_page'][index]
        h_name = re.search(r'fr/(.*?).fr', hostel_url).group(1)
        h_name = h_name.replace('-', ' ')
        
        hostels_list['hotel_name'][index] = h_name
    

hostels_list = hostels_list.dropna().reset_index().drop(columns=['index'])
hostels_list = hostels_list.applymap(html_unicorn)

## Create lat / long columns

hostels_list['latitude'] = np.nan
hostels_list['longitude'] = np.nan

## Clear adress 

def clear_adress(sentence):
    adress = sentence
    try :
        todrop = re.findall(r', (?:...) arr.', adress)[0]
        adress = adress.replace(todrop, '')
    except:
        pass

    try :
        commas_index = re.search(r'[0-9]+(.*?) ', adress).end()-2
        adress = adress[:commas_index] + adress[commas_index+1:]
    except:
        pass

    return adress
    

#pandas shup up
pd.options.mode.chained_assignment = None 

for item in range(len(hostels_list)):
    address = clear_adress(hostels_list['adress'][item])

    url = 'https://nominatim.openstreetmap.org/search/' + urllib.parse.quote(address) +'?format=json'

    response = requests.get(url).json()
    try:
        hostels_list['latitude'][item] = response[0]["lat"]
        hostels_list['longitude'][item] = response[0]["lon"]
    except:
        pass

    print(f'{item}/{len(hostels_list)}')

hostels_list.head()
hostels_list = hostels_list.dropna(subset=['latitude'])

hostels_list.to_csv(r'src/hostels_list_located.csv', index=False)
