==> 09_push_s3.ipynb <==
# Push on AWS S3

import boto3
import logging
import pandas as pd

# gitignore file security AWS password access
# rootkey.csv create in webconsole AWS -> MYNAME -> Info. Id. Secu. -> IAM -> Access Key 
rootkey = pd.read_csv('src/rootkey.csv',header=None, sep="=").T
rootkey.columns = rootkey.iloc[0]
rootkey = rootkey.drop([0])

# Sensitive informations Account
AWSAccessKeyId = rootkey['AWSAccessKeyId'].iloc[0]
AWSSecretKey = rootkey['AWSSecretKey'].iloc[0]

# Open session 
session = boto3.Session(aws_access_key_id=AWSAccessKeyId, 
                        aws_secret_access_key=AWSSecretKey)

# Run session
s3 = session.resource("s3")

# Upload file 
bucket_name = 'compartiment-bucket-s3-jedha-rg'
path = 'src/s3_booking_weather_dataset.csv'
name = 'booking_weather_dataset.csv'
result = s3.Bucket(bucket_name).upload_file(path,name)

# It's OK ?
print(f'Did he encounter a problem ? : {result}')
